refactor: address architecture review, decompose protocol, consolidate workers, and add throughput & anti-snubbing peer scorers
