+++ aura-core/src/peer_registry/mod.rs
@@ -1,2 +1,4 @@
 pub mod logic;
+pub mod scorer;
 pub use logic::*;
+pub use scorer::*;
