+++ aura-core/src/task/mod.rs
@@ -1,2 +1,5 @@
 pub mod logic;
+pub mod extension;
+
 pub use logic::*;
+pub use extension::*;
