+++ aura-core/src/worker/bittorrent/mod.rs
@@ -4,6 +4,8 @@ pub mod protocol;
 #[cfg(test)]
 mod tests;
 pub mod worker;
+pub mod task;
+
 pub use protocol::{
     ExtendedHandshake, Handshake, MetadataMessage, PeerCodec, PeerId, PeerMessage, BLOCK_SIZE,
     HANDSHAKE_LEN,
