+++ aura-core/src/worker/bittorrent/protocol/codec.rs
@@ -1,0 +1,44 @@
+use crate::{Error, Result};
+use bytes::{Buf, BytesMut};
+use tokio_util::codec::{Decoder, Encoder};
+use super::messages::PeerMessage;
+
+pub struct PeerCodec;
+
+impl Decoder for PeerCodec {
+    type Item = PeerMessage;
+    type Error = Error;
+
+    fn decode(&mut self, src: &mut BytesMut) -> Result<Option<Self::Item>> {
+        if src.len() < 4 {
+            return Ok(None);
+        }
+
+        let length = u32::from_be_bytes([src[0], src[1], src[2], src[3]]) as usize;
+
+        if length == 0 {
+            src.advance(4);
+            return Ok(Some(PeerMessage::KeepAlive));
+        }
+
+        if src.len() < 4 + length {
+            src.reserve(4 + length - src.len());
+            return Ok(None);
+        }
+
+        src.advance(4);
+        let data = src.split_to(length);
+        let msg = PeerMessage::deserialize(&data)?;
+        Ok(Some(msg))
+    }
+}
+
+impl Encoder<PeerMessage> for PeerCodec {
+    type Error = Error;
+
+    fn encode(&mut self, item: PeerMessage, dst: &mut BytesMut) -> Result<()> {
+        let serialized = item.serialize();
+        dst.extend_from_slice(&serialized);
+        Ok(())
+    }
+}
