+++ aura-core/src/worker/bittorrent/task/logic.rs
@@ -5,6 +5,7 @@ use crate::peer_registry::PeerRegistry;
 use crate::piece_picker::PiecePicker;
 use crate::torrent::Torrent;
 use crate::tracker::Peer;
+use crate::task::TaskExtension;
 use crate::{Error, InfoHash, Result, TaskId};
 use std::sync::Arc;
 use tokio::sync::{mpsc, Mutex};
@@ -199,3 +200,15 @@ impl BtTask {
         Ok(())
     }
 }
+
+impl TaskExtension for BtTask {
+    fn as_any(&self) -> &dyn std::any::Any {
+        self
+    }
+
+    fn as_any_arc(self: Arc<Self>) -> Arc<dyn std::any::Any + Send + Sync> {
+        self
+    }
+}
+
+
