+++ aura-core/src/worker/mod.rs
@@ -1,3 +1,4 @@
+pub mod bittorrent;
 pub mod builder;
 pub mod ftp;
 pub mod http;
